organise icons into folder
